#96 - 컴포넌트 스캔, 엔티티 스캔 설정 등
* API 모듈, 어드민 모듈이 코어 모듈의 패키지를 컴포넌트 스캔, 엔티티 스캔을 할 수 있도록 설정
* 포트 관련 application.yml 수정
* devtools 외 필요 라이브러리 추가
