--- module-admin/src/main/java/dev/be/ModuleAdminApplication.java
@@ -1,13 +1,0 @@
-package dev.be;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ModuleAdminApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ModuleAdminApplication.class, args);
-    }
-
-}
