+++ module-admin/src/main/resources/application.yml
@@ -102,7 +102,7 @@ spring:
   jpa:
     hibernate:
       generate-ddl: true
-      ddl-auto: create
+      ddl-auto: update
       use-new-id-generator-mappings: false # hibernate_sequence false
     properties:
       hibernate:
