--- module-api/src/main/java/dev/be/ModuleApiApplication.java
@@ -1,13 +1,0 @@
-package dev.be;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ModuleApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ModuleApiApplication.class, args);
-    }
-
-}
