+++ module-api/src/main/java/dev/be/moduleapi/ModuleApiApplication.java
@@ -1,0 +1,21 @@
+package dev.be.moduleapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication(
+        scanBasePackages = {"dev.be.moduleapi", "dev.be.modulecore"}
+)
+@EntityScan("dev.be.modulecore.domain")
+@EnableJpaRepositories(basePackages = {"dev.be.modulecore.repositories"})
+@ConfigurationPropertiesScan
+public class ModuleApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ModuleApiApplication.class, args);
+    }
+
+}
